+++ components/archive/category/src/main/java/com/flipperdevices/archive/category/di/CategoryComponent.kt
@@ -2,6 +2,7 @@ package com.flipperdevices.archive.category.di
 
 import com.flipperdevices.archive.category.fragments.CategoryFragment
 import com.flipperdevices.archive.category.viewmodels.CategoryViewModel
+import com.flipperdevices.archive.category.viewmodels.DeleteViewModel
 import com.flipperdevices.core.di.AppGraph
 import com.squareup.anvil.annotations.ContributesTo
 
@@ -9,4 +10,5 @@ import com.squareup.anvil.annotations.ContributesTo
 interface CategoryComponent {
     fun inject(fragment: CategoryFragment)
     fun inject(viewModel: CategoryViewModel)
+    fun inject(viewModel: DeleteViewModel)
 }