+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 # 1.1 - In progress
 
+- [Feature] Now you can delete or restore all keys from trash
 - [Feature] Add recover support to deleted keys (on keyscreen)